backend/requirements.txt
fastapi
uvicorn
python-arango
python-dotenv
networkx
joblib
numpy
jose
passlib
# smtplib is built into Python and does not need to be installed